Try to do replace-with-noise in forced photometry
This won't help much in CCD forced photometry, as we still don't have
HeavyFootprints there to let us really deblend things, but we might as
well replace sources that aren't in the same deblend family.  In coadd
forced photometry, this is now definitely the right thing to do, as we
can use the HeavyFootprint from consistent single-band deblending
here.