Don't attempt to remove patch overlaps in coadd forced photometry
Now that we're trying to use deblended HeavyFootprints here, we
can't afford to have parent footprints rejected as being in the
overlap region if their children might not be rejected, because
the ReplaceWithNoiseTask requires all sources to have a parent.

In order to change this back, we'd need more intelligent handling
of the overlaps.  But once this is loaded into the database,
we should be able to join with the main measurement catalog and
just use the "primary" flags there to handle overlaps.