Set mask plane BRIGHT_OBJECT as per HSC-1342
assembleCoadd.py reads an externally provided specification of
bright object masks (via the butler type "brightObjectMask" which is added in obs_subaru on this ticket) and sets a mask
plane in the exposure, by default BRIGHT_OBJECT.  This is intended to allow us to mask bright stars and other troublesome objects for LSS and other studies.

obs/subaru/config/assembleCoadd.py sets flag bits in sources that fall within this mask.